feat: refactor WebSocket handling and move models to separate files
